+++ packages/e2e-tests/tests/common/api-helpers/collection-helper.ts
@@ -9,7 +9,6 @@ import type {
 } from "@rarible/api-client"
 import type { Collections } from "@rarible/api-client/build/models"
 
-
 export async function getCollectionById(sdk: IRaribleSdk, collectionId: string): Promise<Collection> {
 	const collection = await retry(15, 3000, async () => {
 		return await sdk.apis.collection.getCollectionById({
@@ -86,9 +85,11 @@ export async function verifyCollectionsByBlockchain(collections: Collections, bl
 	})
 }
 
-export async function verifyCollectionsContainsCollection(collections: Collections,
-	collectionId: string): Promise<void> {
-	expect(collections.collections.map(c => c.id as string)).toContain(collectionId)
+export async function verifyCollectionsContainsCollection(
+	collections: Collections,
+	collectionId: string
+): Promise<void> {
+	expect(collections.collections.map(c => c.id.toLowerCase())).toContain(collectionId.toLowerCase())
 }
 
 export async function verifyCollectionsOwner(collections: Collections, owner: string): Promise<void> {