+++ packages/e2e-tests/tests/common/config.ts
@@ -1,19 +1,9 @@
-import type { Blockchain } from "@rarible/api-client"
 import { e2eTestsConfig } from "./config/e2e"
 import { developmentTestsConfig } from "./config/development"
 
 export const testsConfig = {
-	blockchain: getBlockchain(),
 	env: process.env.TEST_ENV,
 	variables: fetchEnv(process.env.TEST_ENV),
-}
-
-function getBlockchain() {
-	let blockchains = process.env.BLOCKCHAIN?.split(",")
-	if (blockchains === undefined || blockchains.includes("ALL")) {
-		blockchains = ["ETHEREUM", "TEZOS", "FLOW", "SOLANA"]
-	}
-	return blockchains as Blockchain[]
 }
 
 function fetchEnv(env?: string) {