+++ packages/e2e-tests/tests/common/logger.ts
@@ -1,0 +1,9 @@
+const logsEnabled = process.env.LOGGER_ENABLED === "1"
+
+export class Logger {
+	static log(...args: any[]) {
+		if (logsEnabled) {
+			console.log(...args)
+		}
+	}
+}