+++ packages/e2e-tests/tests/tezos/smoke/order/mint-sell-sellupdate-buy.test.ts
@@ -57,7 +57,7 @@ function suites(): {
 				}
 			},
 			updateSellRequest: {
-				price: "0.01",
+				price: "0.04",
 			},
 		},
 		{