+++ packages/connector-helper/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.11-beta.21](https://github.com/rarible/sdk/compare/v0.9.11-beta.20...v0.9.11-beta.21) (2022-07-20)
+
+
+### Bug Fixes
+
+* royalties ([75846b4](https://github.com/rarible/sdk/commit/75846b482685d15ede793326d7118d71ecceb8e5))
+
+
+
+
+
+## [0.9.11-beta.20](https://github.com/rarible/sdk/compare/v0.9.11-beta.19...v0.9.11-beta.20) (2022-07-19)
+
+**Note:** Version bump only for package @rarible/connector-helper
+
+
+
+
+
+## [0.9.11-beta.19](https://github.com/rarible/sdk/compare/v0.9.11-beta.18...v0.9.11-beta.19) (2022-07-18)
+
+**Note:** Version bump only for package @rarible/connector-helper
+
+
+
+
+
+## [0.9.11-beta.18](https://github.com/rarible/sdk/compare/v0.9.11-beta.17...v0.9.11-beta.18) (2022-07-18)
+
+
+### Bug Fixes
+
+* tezos royalties ([a4b7f6b](https://github.com/rarible/sdk/commit/a4b7f6bc27649928b0dd3d4e2e5feb7e89931367))
+
+
+
+
+
 ## [0.9.11-beta.17](https://github.com/rarible/sdk/compare/v0.9.11-beta.16...v0.9.11-beta.17) (2022-07-15)
 
 **Note:** Version bump only for package @rarible/connector-helper